+++ AddOns/XIV_Databar_Continued/CHANGELOG.md
@@ -1,9 +1,14 @@
 # XIV_Databar Continued
 
-## [3.7.6](https://github.com/ZelionGG/XIV_Databar-Continued/releases/tag/3.7.6) (2025-12-03)
+## [3.8](https://github.com/ZelionGG/XIV_Databar-Continued/releases/tag/3.8) (2025-12-04)
 
-[Full Changelog](https://github.com/ZelionGG/XIV_Databar-Continued/compare/v3.7.5...v3.7.6) [Previous Releases](https://github.com/ZelionGG/XIV_Databar-Continued/releases)
+[Full Changelog](https://github.com/ZelionGG/XIV_Databar-Continued/compare/v3.7.6...v3.8) [Previous Releases](https://github.com/ZelionGG/XIV_Databar-Continued/releases)
 
 ### _Retail :_
 
-- Updated for patch **11.2.7**.
+- Added Housing Dashboard button to **Micromenu** module (enabled by default).
+- Disabled all **Micromenu** buttons in combat to avoid Taint errors.
+
+### _Improvements :_
+
+- Improved French locale.